Bird categories page
Completed basic setup of the bird categories page. Downloaded all the images and organized them into image files. Created a flexbox for the image links with a centered title underneath. Added responsive styling to scale images with the browser size and change the number of "columns" in the flexbox. Fixed the navigation link references on lower pages and added a base url to all file heads. Added a universal color to body and p.