Prototype first image gallery
Done on 08 May 2020. Set up enough of the site to show JoAnne a prototype. Created an image gallery for the brazil bird images. Uses php to grab all the image files in the folder and display them in a grid. Styled the grid to auto fit the images, and crop the images into the grid spaces. Fixed the navigation and image links to work with the new base link. Renamed the categories class.