Meeting feedback adjustments
Met with JoAnne to show her the site prototype, and make adjustments bases on the topics I brought up and her feedback. Decided to go with one gallery for all the categories, instead of seperate one for birds, wildlife, etc. Changed the folder organization and navigation links to reflect the change. Added a google font to the head, with fallback fonts. Lightened the active nav's color, and added an on hover fade to the navigation links.  Added a centered title above the gallery grid. Added more white space to the sides of the gallery, and lessened the space between images.