Set up image modal
Set up the basic parts of the image modal in the image galleries. Edited php to include the modal functions. Added the modal container with controls, a close button, and php code to grab the images. Added a modal script to display and close the modal, and scroll through the gallery images in the modal. Added initial styling for the modal, buttons, controls, and mostly centered images. Still needs additional styling, proper image centering, overlay buttons, preventing scrolling, and animations.