Improved Modal
Made changes to improve the syling and functionality of the modal. Modified the modal functions to prevent the gallery page from scrolling while the modal is open, and return to the scroll point when the modal closes. Added full screen overlay controls to the modal, with nicer looking arrows that change shade on overlay hover. Moved around modal elements in the z-index to prevent the image overlapping the controls. Edited the footer to always stay at the bottom of the page and never overlap other page elements.