Mobile Nav
Implemented the mobile navigation and styled the mobile site. Created a navigation script for the collaspible mobile icon. Added and styled a navigation icon to the navigation bar that opens and closes the links, and positioned it on the page. Styled the mobile navigation bar to block display instead of grid, and lessened most padding. Added minor navigation padding to match up with other elements. Edited the navigation link styling to not effect the title and navigation icon. Added a hover for the categories page. Added a transition ease to the arrow controls on the gallery modal.