Galleries Rework
Reworked the galleries sucessfully this time to a responsive grid with embedded grids, and reverted images in categories and gallery back to being cropped. Reworked the html structure to seperate the image and text in the grid. Images get cropped to fit the link, and the text changes color when you hover over the image or the text.