Slideshow + Gallery Update
Added new slideshow and gallery images to the folder system. Not all galleries have all their images. Styled the slideshow images to be centered in a container. Added the existing slideshow script to the homepage, and re-added the animation styling. Added the images and links to the new galleries to the galleries page, and changed the way the gallery images height is displayed to prevent tall category images. Added some responsive height edits for certain viewports for the slideshow and gallery containers, and the categories images, mostly to prevent overly cropped images. Deleted some old commented code.