Modal Caption Set Up
Set up inital captions for the gallery modal images. Created text files with placeholder information for each brazil bird. Added php to grab the files from the directory and create caption divs. Uncommented script code to switch the caption with the images. Started styling the caption text, trying to get it to appear on hover over the image.