Modal Rework for Captions
Reworked the modal to have a different structure so the captions can sit at the bottom of each image dynamically. The issue was the modal container could not get the height of absolutely positioned images with pure CSS. Wrote new javascript code to calculate the heights of each image container based on their respective image's dimensions. Reworked the gallery php to find individual text files and read them out into a container div with their corresponding image, instead of all the captions listed after all the images. Styled the container for centering instead of the images.