Modal Bugfixing
Fixed some issues with the modal rework, particularly with browsers other than Chrome. Now calculates the image container height when the modal is opened, instead of when switching images, as Firefox and Edge weren't displaying the calculated height the first time the image was viewed. Now sets the image capacity to 0 when closing, so it doesn't transition from the previous image when opening the modal again. Recalculates the next image's height when switching to get Firefox and Edge to display the actual resized height instead of the image's base height. Added -webkit- selectors to work with more browsers. Added will-change on hover to make transitions smoother. Commented new script code. Fixed an issue where the first image opened would transition from the first image in the gallery, since the image index is set to 1. Issue with Edge where the first time the modal is opened, the images shift when the browser calculates the new height (but no one uses Edge so it's not a big deal).