Modal Captions Functional
Finished the reworks for the modal caption text. Reworked how the overlay controls worked from two divs placed above the images, which blocked the images from being hoverable. Now a function runs when the image container is clicked, which calculates if the mouse clicked to the left or right of the painting and goes to the next or previous image. Also makes all image containers except the current image not respond to pointer events, so only the current image's caption can appear and can be clicked. Styled the caption container and text. Changed the modal gallery arrows to have the z index instead of their containers so they appear over the images when overlapping. Deleted old commented code. Added explanation comments to the new script code.