Ordered Galleries
Odered the images and caption files of the complete galleries. Since the file order depends on the server, which does it alphabetically, each image has a number at the beginning of its name to denote it's order in the gallery. Deleted some images the client doesn't want on the site. Added new images and caption files for them, and compressed them. Changed the gallery category images to the first image in each complete gallery. Added the new complete galapagos birds gallery and made caption text files for each image.