Bugfixes and Edits
Fixed a bug on short wide monitors where the About page image was overflowing past the content container. Fixed a caption with unecessary numerators. Commented out the unfinished gallery categories in Galleries. Added alt image text to the site. Also validated the site ownership with google and submitted the sitemap.