Fixes
Made some fixes that the client asked for. Added the galapagos gallery, which I forgot I had already completed, and the maui gallery with it's new compressed images and caption files. Fixed some images with artifacting in the background due to too much compression. Reduced the number of slideshow images and renumbered the remaining ones. Changed the Facebook share image for the about and contact page to be JoAnne instead of a bird. Also completed the guide on editing the site.