Texas Gallery + Updates
Created a gallery page for Texas birds. Created a php file to call and display the new images in the folder. Compressed the images and added the text files JoAnne created. Added a link to the new gallery on the categories page, with the new image for the icon. Added two new images to the slideshow. Edited the text on the About page. Updated the sitemap.