New Images cont.
Optimized the rest of the new image files. Created php files for all the new galleries. Added the new galleries to the galleries page. Edited some names (texas and alaska) in their gallery pages and on the galleries page. Edited some image names to have a 0 in front of a single digit if it was missing, so it will properly order on bluehost. Also realized California had new images so did everything with those too. Still need to generate captions for all images and test the new pages on the live site.