Gallery Tabs
Added a grid container for gallery tab links, spent a while remembering how repeat works. Created and styled the css. Created the colombia 2024 gallery file and added the gallery tabs to the old colombia and new files. Added 2020 to old gallery meta tags. Tested in apache and uploaded to website. Gallery tabs set max size to 15%, should be enough but may need to widen for the other galleries that need gallery tab integration.